chapters/chapter_7/pyproject.toml
[tool.poetry]
name = "chapter_7"
version = "0.1.0"
description = ""
authors = ["Daniel van Flymen <dvf@candidco.com>"]

[tool.poetry.dependencies]
python = "^3.8"
aiohttp = "^3.6.2"
pynacl = "^1.3.0"
structlog = "^20.1.0"
colorama = "^0.4.3"
marshmallow = "^3.6.0"
marshmallow-oneofschema = "^2.0.1"

[tool.poetry.dev-dependencies]
pytest = "^6"
black = "^20.8b1"
pytest-asyncio = "*"
pytest-mock = "^3.3.1"
pytest-faker = "^2.0.0"
ipython = "^7.18.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
